Update Q1_Task1.py
Updated the CSV file name to point to the Assignment 2.zip as downloaded so the user is not required to rename to CSV.zip. Noting we also achieved the function of task 1 with the 'csv' open source library. However, pandas appears more flexible if future options were desired and the csv library was not any more efficient than pandas library at processing the data.